requirements.txt
Django~=3.2.7
django-cors-headers~=3.4.0
djangorestframework~=3.11.2
django-environ==0.9.0
Pillow==8.1.2
djangorestframework-simplejwt==4.5.0
PyJWT==1.7.1
dj-rest-auth==2.1.3
django-allauth==0.44.0
psycopg2-binary==2.9.1
requests==2.28.1